+++ test/smoke/data/runtime/src/wrong-validate-in-scope.js
@@ -1,0 +1,13 @@
+// @typecheck
+
+function Test() {}
+
+/**
+ * @param {Test} data
+ * @returns {null}
+ */
+function testFunction(data) {
+    return null;
+}
+
+testFunction('wrong data');